+++ src/TestUtils/Generator/TestDataContainerGenerator.php
@@ -3,7 +3,6 @@ declare(strict_types=1);
 
 namespace Raptor\TestUtils\Generator;
 
-use Memcache;
 use Raptor\TestUtils\DataLoader\DirectoryDataLoader;
 use Raptor\TestUtils\DataProcessor\Type\Type;
 use Raptor\TestUtils\Exceptions\DataDirectoryNotFoundException;